Update paths that broke when micro_ed was moved into scripts.
